refactor(cli): extract runtime history handler
